+++ autoconfigure/pom.xml
@@ -4,15 +4,27 @@
     <artifactId>parent</artifactId>
     <groupId>com.google.cloud.trace.adapters.zipkin</groupId>
     <version>0.2.1-SNAPSHOT</version>
+    <relativePath>../pom.xml</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
   <packaging>jar</packaging>
   <version>0.2.1-SNAPSHOT</version>
-  <artifactId>autoconfigure-storage</artifactId>
-  <name>Autoconfigure Stackdriver Zipkin Storage</name>
+  <artifactId>autoconfigure</artifactId>
+  <name>Autoconfigure Stackdriver Zipkin</name>
 
   <dependencies>
+    <dependency>
+      <groupId>io.zipkin.java</groupId>
+      <artifactId>zipkin-server</artifactId>
+      <version>${zipkin.version}</version>
+      <exclusions>
+        <exclusion>
+          <groupId>org.springframework.boot</groupId>
+          <artifactId>spring-boot-starter-tomcat</artifactId>
+        </exclusion>
+      </exclusions>
+    </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-autoconfigure</artifactId>
@@ -29,28 +41,14 @@
     <dependency>
       <groupId>io.zipkin.java</groupId>
       <artifactId>zipkin</artifactId>
-      <version>${zipkin.version}</version>
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
       <artifactId>storage</artifactId>
-      <version>0.2.1-SNAPSHOT</version>
     </dependency>
     <dependency>
       <groupId>com.google.auth</groupId>
       <artifactId>google-auth-library-oauth2-http</artifactId>
-      <version>0.4.0</version>
-      <exclusions>
-        <exclusion>
-          <groupId>com.google.guava</groupId>
-          <artifactId>guava-jdk5</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-actuator</artifactId>
     </dependency>
   </dependencies>
-
 </project>
