+++ collector/pom.xml
@@ -4,6 +4,7 @@
     <artifactId>parent</artifactId>
     <groupId>com.google.cloud.trace.adapters.zipkin</groupId>
     <version>0.2.1-SNAPSHOT</version>
+    <relativePath>../pom.xml</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <packaging>jar</packaging>
@@ -21,29 +22,8 @@
   -->
   <dependencies>
     <dependency>
-      <groupId>io.zipkin.java</groupId>
-      <artifactId>zipkin-server</artifactId>
-      <version>${zipkin.version}</version>
-      <exclusions>
-        <exclusion>
-          <groupId>org.springframework.boot</groupId>
-          <artifactId>spring-boot-starter-tomcat</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-jetty</artifactId>
-    </dependency>
-    <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>autoconfigure-storage</artifactId>
-      <version>0.2.1-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>io.netty</groupId>
-      <artifactId>netty-tcnative-boringssl-static</artifactId>
-      <version>1.1.33.Fork23</version>
+      <artifactId>starter</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -55,7 +35,6 @@
       <artifactId>zipkin</artifactId>
       <scope>test</scope>
       <type>test-jar</type>
-      <version>${zipkin.version}</version>
     </dependency>
   </dependencies>
 <build>
