+++ translation/pom.xml
@@ -4,6 +4,7 @@
     <artifactId>parent</artifactId>
     <groupId>com.google.cloud.trace.adapters.zipkin</groupId>
     <version>0.2.1-SNAPSHOT</version>
+    <relativePath>../pom.xml</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
@@ -16,17 +17,14 @@
     <dependency>
       <groupId>io.zipkin.java</groupId>
       <artifactId>zipkin</artifactId>
-      <version>${zipkin.version}</version>
     </dependency>
     <dependency>
       <groupId>com.google.api.grpc</groupId>
       <artifactId>grpc-google-devtools-cloudtrace-v1</artifactId>
-      <version>0.1.0</version>
     </dependency>
     <dependency>
       <groupId>io.grpc</groupId>
       <artifactId>grpc-all</artifactId>
-      <version>${grpc.version}</version>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
